+++ src/Gameplay/Entities/PacMan.java
@@ -8,7 +8,7 @@ import java.util.*;
  * Gestion du personnage Pac-Man
  */
 public class PacMan extends DynamicEntity {
-    public PacMan(List<Double> speed, List<Double> position, List<Double> dimensions, String fileName){
+    public PacMan(List<Integer> speed, List<Integer> position, List<Integer> dimensions, String fileName){
         super(speed, position, dimensions, fileName);
     }
 }
